requirements.txt

easyocr
opencv-python
matplotlib 
numpy
pandas 

pytesseract 
Pillow

spacy

IPython
